fix: prevent album detail tap-through
